// src/domain/model/customer.ts
export interface CustomerInformation {
  customer_name: string;
  account_stage: string;
  deal_stage: string;
  mrr: number;
  arr: number;
  discount: number;
  natural_margin: number;
  gross_margin: number;
  deal_owner: string;
  primary_assigned: string;
  created_date: string;
  last_modified_by: string;
}